🚚 rename repository correctly
